Move rate limiter to redis package
