+++ spring-cloud-gateway-core/src/main/java/org/springframework/cloud/gateway/filter/redis/ratelimit/RateLimiter.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.springframework.cloud.gateway.filter.ratelimit;
+package org.springframework.cloud.gateway.filter.redis.ratelimit;
 
 import java.util.Collections;
 import java.util.Map;
