+++ spring-cloud-gateway-core/src/test/java/org/springframework/cloud/gateway/filter/redis/ratelimit/PrincipalNameKeyResolverIntegrationTests.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.springframework.cloud.gateway.filter.ratelimit;
+package org.springframework.cloud.gateway.filter.redis.ratelimit;
 
 import java.security.Principal;
 import java.util.Collections;
