+++ spring-cloud-gateway-core/src/test/java/org/springframework/cloud/gateway/filter/redis/ratelimit/RedisRateLimiterConfigTests.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.springframework.cloud.gateway.filter.ratelimit;
+package org.springframework.cloud.gateway.filter.redis.ratelimit;
 
 import org.junit.Before;
 import org.junit.Test;
@@ -26,6 +26,8 @@ import org.springframework.boot.autoconfigure.EnableAutoConfiguration;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.cloud.gateway.route.Route;
 import org.springframework.cloud.gateway.route.RouteLocator;
+import org.springframework.cloud.gateway.route.builder.RouteLocatorBuilder;
+import org.springframework.context.annotation.Bean;
 import org.springframework.test.annotation.DirtiesContext;
 import org.springframework.test.context.ActiveProfiles;
 import org.springframework.test.context.junit4.SpringRunner;
@@ -38,8 +40,8 @@ import static org.assertj.core.api.Assertions.assertThat;
 @RunWith(SpringRunner.class)
 @SpringBootTest
 @DirtiesContext
-@ActiveProfiles("redis-rate-limiter-default-config")
-public class RedisRateLimiterDefaultFilterConfigTests {
+@ActiveProfiles("redis-rate-limiter-config")
+public class RedisRateLimiterConfigTests {
 
 	@Autowired
 	private RedisRateLimiter rateLimiter;
@@ -54,14 +56,31 @@ public class RedisRateLimiterDefaultFilterConfigTests {
 	}
 
 	@Test
-	public void redisRateConfiguredFromEnvironmentDefaultFilters() {
-		String routeId = "redis_rate_limiter_config_default_test";
-		RedisRateLimiter.Config config = rateLimiter.loadConfiguration(routeId);
-		assertConfigAndRoute(routeId, 70, 80, config);
+	public void redisRateConfiguredFromEnvironment() {
+		assertFilter("redis_rate_limiter_config_test", 10, 20, false);
 	}
 
-	private void assertConfigAndRoute(String key, int replenishRate, int burstCapacity,
-			RedisRateLimiter.Config config) {
+	@Test
+	public void redisRateConfiguredFromJavaAPI() {
+		assertFilter("custom_redis_rate_limiter", 20, 40, false);
+	}
+
+	@Test
+	public void redisRateConfiguredFromJavaAPIDirectBean() {
+		assertFilter("alt_custom_redis_rate_limiter", 30, 60, true);
+	}
+
+	private void assertFilter(String key, int replenishRate, int burstCapacity,
+			boolean useDefaultConfig) {
+		RedisRateLimiter.Config config;
+
+		if (useDefaultConfig) {
+			config = rateLimiter.getDefaultConfig();
+		}
+		else {
+			assertThat(rateLimiter.getConfig()).containsKey(key);
+			config = rateLimiter.getConfig().get(key);
+		}
 		assertThat(config).isNotNull();
 		assertThat(config.getReplenishRate()).isEqualTo(replenishRate);
 		assertThat(config.getBurstCapacity()).isEqualTo(burstCapacity);
@@ -69,12 +88,33 @@ public class RedisRateLimiterDefaultFilterConfigTests {
 		Route route = routeLocator.getRoutes().filter(r -> r.getId().equals(key)).next()
 				.block();
 		assertThat(route).isNotNull();
-		assertThat(route.getFilters()).isNotEmpty();
+		assertThat(route.getFilters()).hasSize(1);
 	}
 
 	@EnableAutoConfiguration
 	@SpringBootConfiguration
 	public static class TestConfig {
+
+		@Bean
+		public RouteLocator testRouteLocator(RouteLocatorBuilder builder) {
+			return builder.routes().route("custom_redis_rate_limiter",
+					r -> r.path("/custom").filters(f -> f.requestRateLimiter()
+							.rateLimiter(RedisRateLimiter.class,
+									rl -> rl.setBurstCapacity(40).setReplenishRate(20))
+							.and()).uri("http://localhost"))
+					.route("alt_custom_redis_rate_limiter",
+							r -> r.path("/custom")
+									.filters(f -> f.requestRateLimiter(
+											c -> c.setRateLimiter(myRateLimiter())))
+									.uri("http://localhost"))
+					.build();
+
+		}
+
+		@Bean
+		public RedisRateLimiter myRateLimiter() {
+			return new RedisRateLimiter(30, 60);
+		}
 
 	}
