+++ spring-cloud-gateway-core/src/test/java/org/springframework/cloud/gateway/test/AdhocTestSuite.java
@@ -92,9 +92,9 @@ import static org.junit.Assume.assumeThat;
 		org.springframework.cloud.gateway.filter.headers.NonStandardHeadersInResponseTests.class,
 		org.springframework.cloud.gateway.filter.headers.RemoveHopByHopHeadersFilterTests.class,
 		org.springframework.cloud.gateway.filter.WebsocketRoutingFilterTests.class,
-		org.springframework.cloud.gateway.filter.ratelimit.PrincipalNameKeyResolverIntegrationTests.class,
-		org.springframework.cloud.gateway.filter.ratelimit.RedisRateLimiterConfigTests.class,
-		org.springframework.cloud.gateway.filter.ratelimit.RedisRateLimiterTests.class,
+		org.springframework.cloud.gateway.filter.redis.ratelimit.PrincipalNameKeyResolverIntegrationTests.class,
+		org.springframework.cloud.gateway.filter.redis.ratelimit.RedisRateLimiterConfigTests.class,
+		org.springframework.cloud.gateway.filter.redis.ratelimit.RedisRateLimiterTests.class,
 		org.springframework.cloud.gateway.filter.LoadBalancerClientFilterTests.class,
 		org.springframework.cloud.gateway.filter.NettyRoutingFilterIntegrationTests.class,
 		GatewayMetricsFilterTests.class,
